Refactor auto layout
